Fix clipboard issues on Win11
The clipboard handling code uses OpenClipboard() to actually grab clipboard
handle. The function can fail if other process invokes it in same time.

If it fails, the Windows client does nothing. Clipboard content in question
is silently dropped. This works somewhat ok on Win10, but on Win11 for
some reason OpenClipboard() almost never succeeds from first time.

This is a known problem, and it seems like there is no other way to solve
this except for aggressive clipboard grabbing. This commit implements
exactly that.